+++ types/erp.ts
@@ -196,6 +196,8 @@ export enum FormatoDocumentoAuxiliarFiscal {
 
 export type PermissionCode =
     | 'AUDITORIA_CONSULTAR'
+    | 'ATIVIDADES_CONSULTAR'
+    | 'ATIVIDADES_GERENCIAR'
     | 'ADMINISTRACAO_CONSULTAR'
     | 'ADMINISTRACAO_GERENCIAR'
     | 'SEGURANCA_USUARIOS_CONSULTAR'